Fix keybinding not being available to override through configuration settings.
In the function changemapping() the linked list of keybindings is traversed in
a way the last element of the list (tail) wont be processed to rebind.
The loop iterates over all elements of the linked list until an element
which has none _next_ to it, ie: the last one, efectibly skipping the last element
in the list.

The changes fix this by iterating over all elements until finding a NULL element.

Patch by desyncr via GitHub:
https://github.com/desyncr/vimprobable/commit/050a96f849e7b14ba2e2629268ee559c5afe6856